better variation resizing
