+++ src/components/ParameterView.tsx
@@ -1,61 +1,29 @@
 import { ExtraParams, PatternParam } from "@/src/types/PatternParams";
-import { Box, Button, Divider, HStack, Text, VStack } from "@chakra-ui/react";
-import { memo, useEffect, useState } from "react";
+import { Box, Button, Divider, HStack, Text } from "@chakra-ui/react";
+import { useEffect, useState } from "react";
 import { BsCaretDown, BsCaretUp } from "react-icons/bs";
-import Variation from "@/src/types/Variations/Variation";
-import VariationGraph from "@/src/components/VariationGraph/VariationGraph";
-import {
-  DragDropContext,
-  Draggable,
-  Droppable,
-  OnDragEndResponder,
-} from "@hello-pangea/dnd";
-import { reorder } from "@/src/utils/algorithm";
 import Block from "@/src/types/Block";
-import { action } from "mobx";
-import VariationBound from "@/src/components/VariationBound";
 import NewVariationButtons from "@/src/components/NewVariationButtons";
+import ParameterVariations from "@/src/components/ParameterVariations";
+import { observer } from "mobx-react-lite";
 
 type ParameterProps = {
   uniformName: string;
   patternParam: PatternParam;
   block: Block<ExtraParams>;
-  variations: Variation[];
-  width: number;
 };
 
-export default memo(function ParameterView({
+export default observer(function ParameterView({
   uniformName,
   patternParam,
   block,
-  variations,
-  width,
 }: ParameterProps) {
+  const variations = block.parameterVariations[uniformName] ?? [];
   const [isExpanded, setExpanded] = useState(false);
 
   // only expand once we are on the client, otherwise strange errors
   useEffect(() => setExpanded(true), [setExpanded]);
 
-  const domain: [number, number] = [0, 1];
-  for (const variation of variations) {
-    const [min, max] = variation.computeDomain();
-    domain[0] = Math.min(domain[0], min);
-    domain[1] = Math.max(domain[1], max);
-  }
-
-  const onDragEnd: OnDragEndResponder = action((result) => {
-    // dropped outside the list, delete
-    if (!result.destination) {
-      block.removeVariation(uniformName, variations[result.source.index]);
-      return;
-    }
-
-    block.parameterVariations[uniformName] = reorder(
-      variations,
-      result.source.index,
-      result.destination.index
-    );
-  });
   return (
     <>
       <Divider borderWidth={1} borderColor="gray.500" />
@@ -92,72 +60,13 @@ export default memo(function ParameterView({
             <NewVariationButtons uniformName={uniformName} block={block} />
           </HStack>
         ) : (
-          <DragDropContext onDragEnd={onDragEnd}>
-            <Droppable
-              droppableId={block.id + uniformName}
-              direction="horizontal"
-            >
-              {(provided, snapshot) => (
-                <HStack
-                  ref={provided.innerRef}
-                  {...provided.droppableProps}
-                  width="100%"
-                  justify="start"
-                  spacing={0}
-                >
-                  {variations.map((variation, index) => (
-                    <>
-                      <Draggable
-                        key={variation.id}
-                        draggableId={variation.id}
-                        index={index}
-                      >
-                        {(provided, snapshot) => (
-                          <VStack
-                            ref={provided.innerRef}
-                            {...provided.draggableProps}
-                            {...provided.dragHandleProps}
-                          >
-                            <VariationGraph
-                              uniformName={uniformName}
-                              variation={variation}
-                              width={
-                                variation.duration < 0
-                                  ? width
-                                  : (variation.duration / block.duration) *
-                                    width
-                              }
-                              domain={domain}
-                              block={block}
-                            />
-                          </VStack>
-                        )}
-                      </Draggable>
-                      {/* <Box width={1} /> */}
-                    </>
-                  ))}
-                  {provided.placeholder}
-                  <NewVariationButtons
-                    uniformName={uniformName}
-                    block={block}
-                  />
-                </HStack>
-              )}
-            </Droppable>
-          </DragDropContext>
+          <ParameterVariations
+            uniformName={uniformName}
+            block={block}
+            variations={variations}
+          />
         ))}
-      {isExpanded && (
-        <HStack alignSelf="flex-start" justify="start" spacing={0} pb={2}>
-          {variations.map((variation) => (
-            <VariationBound
-              key={variation.id}
-              uniformName={uniformName}
-              block={block}
-              variation={variation}
-            />
-          ))}
-        </HStack>
-      )}
+      <Box height={2} />
     </>
   );
 });
