+++ src/components/ParametersList.tsx
@@ -1,9 +1,8 @@
 import Block from "@/src/types/Block";
 import { ExtraParams } from "@/src/types/PatternParams";
 import ParameterView from "@/src/components/ParameterView";
-import { observer } from "mobx-react-lite";
-import { useStore } from "@/src/types/StoreContext";
 import { VStack } from "@chakra-ui/react";
+import { memo } from "react";
 
 const uniformNamesToExclude = ["u_resolution", "u_time", "u_global_time"];
 
@@ -11,12 +10,7 @@ type ParametersListProps = {
   block: Block<ExtraParams>;
 };
 
-export default observer(function ParametersList({
-  block,
-}: ParametersListProps) {
-  const { uiStore } = useStore();
-  const width = uiStore.timeToX(block.duration);
-
+export default memo(function ParametersList({ block }: ParametersListProps) {
   return (
     <VStack spacing={0} width="100%">
       {Object.entries(block.pattern.params).map(([uniformName, patternParam]) =>
@@ -26,8 +20,6 @@ export default observer(function ParametersList({
             uniformName={uniformName}
             patternParam={patternParam}
             block={block}
-            variations={block.parameterVariations[uniformName] ?? []}
-            width={width}
           />
         )
       )}
